% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_installer.R
\name{get_r_installer}
\alias{get_r_installer}
\title{Get R installer}
\usage{
get_r_installer(x)
}
\arguments{
\item{x}{A character vector with one element.}
}
\value{
A character vector.
}
\description{
Get R installer
}
\examples{
x <- "alfa,bravo,charlie,delta"
get_r_installer(x)
}
